+++ scanpipe/models.py
@@ -78,7 +78,6 @@ from rq.job import Job
 from rq.job import JobStatus
 
 from scancodeio import __version__ as scancodeio_version
-from scanpipe import spdx
 from scanpipe import tasks
 
 logger = logging.getLogger(__name__)
@@ -1873,6 +1872,8 @@ class CodebaseResource(
 
     def as_spdx(self):
         """Return this CodebaseResource as an SPDX Package entry."""
+        from scanpipe.pipes import spdx
+
         spdx_license_keys = [license["spdx_license_key"] for license in self.licenses]
         copyrights = [copyright["copyright"] for copyright in self.copyrights]
         holders = [holder["holder"] for holder in self.holders]
@@ -2200,6 +2201,8 @@ class DiscoveredPackage(
 
     def as_spdx(self):
         """Return this DiscoveredPackage as an SPDX Package entry."""
+        from scanpipe.pipes import spdx
+
         checksums = [
             spdx.Checksum(algorithm=algorithm, value=checksum_value)
             for algorithm in ["sha1", "md5"]
@@ -2496,6 +2499,8 @@ class DiscoveredDependency(
 
     def as_spdx(self):
         """Return this Package as an SPDX Package entry."""
+        from scanpipe.pipes import spdx
+
         external_refs = []
 
         if package_url := self.package_url: